requirements.txt
# === PyPI ===
--index-url https://pypi.org/simple

# === PyTorch CUDA wheels ===
--extra-index-url https://download.pytorch.org/whl/cu118

# === Core Libraries ===
numpy==1.24.4
pandas==2.2.1
scipy==1.12.0
scikit-learn==1.4.1.post1
matplotlib==3.8.3
tqdm==4.66.2
PyYAML==6.0

# === MNE for EEG/BCI ===
mne==1.7.0

# === DL Libraries (CUDA 11.8) ===
torch==2.2.2+cu118
torchvision==0.17.2+cu118
torchaudio==2.2.2+cu118

# === TensorFlow / Keras (GPU-safe versions) ===
tensorflow==2.15.0
keras==2.15.0

# === Imbalanced Learning ===
imbalanced-learn==0.12.0
